retrieve coordinate reference system from object

retrieve coordinate reference system from sf or sfc object.

set or replace retrieve coordinate reference system from object

st_crs(x, ...)

# S3 method for sf
st_crs(x, ...)

# S3 method for sfc
st_crs(x, ...)

st_crs(x) <- value

# S3 method for sf
st_crs(x) <- value

# S3 method for sfc
st_crs(x) <- value

Arguments

x

object of class sf or sfc

...

ignored

value

one of (i) character: a valid proj4string (ii) integer, a valid epsg value (numeric), or (iii) a list containing named elements proj4string (character) and/or epsg (integer) with (i) and (ii).

Details

in case a coordinate reference system is replaced, no transformation takes place and a warning is raised to stress this. epsg values are either read from proj4strings that contain +init=epsg:... or set to 4326 in case the proj4string contains +proj=longlat and +datum=WGS84, literally

If both epsg and proj4string are provided, they are assumed to be consistent. In processing them, the epsg code, if not missing valued, is used and the proj4string is derived from it by a call to GDAL (which in turn will call PROJ.4). Warnings are raised when epsg is not consistent with a proj4string that is already present.

Examples

 sfc = st_sfc(st_point(c(0,0)), st_point(c(1,1)))
 sf = st_sf(a = 1:2, geom = sfc)
 st_crs(sf) = 4326
 st_geometry(sf)

#> Geometry set for 2 features 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 1 ymax: 1
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> POINT(0 0) 
#> POINT(1 1) 
#> 

 sfc = st_sfc(st_point(c(0,0)), st_point(c(1,1)))
 st_crs(sfc) = 4326
 sfc

#> Geometry set for 2 features 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 1 ymax: 1
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#> POINT(0 0) 
#> POINT(1 1) 
#>