Check if the columns could be of a coercable type for sf

Check if the columns could be of a coercable type for sf

is_geometry_column(con, x, classes = "")

Arguments

con
database connection

x
inherits data.frame

classes
classes inherited