sf-defunct {sf}	R Documentation

Deprecated functions in sf

Description

These functions are provided for compatibility with older version of sf. They may eventually be completely removed.

Usage

st_read_db(
  conn = NULL,
  table = NULL,
  query = NULL,
  geom_column = NULL,
  EWKB = TRUE,
  ...
)

Arguments

conn
open database connection

table
table name

query
SQL query to select records; see details

geom_column
deprecated. Geometry column name

EWKB
logical; is the WKB of type EWKB? if missing, defaults to TRUE

...
parameter(s) passed on to st_as_sf

Details

The geom_column argument is deprecated. The function will automatically find the geometry type columns. For the RPostgreSQL drivers it will try to cast all the character columns, which can be long for very wide tables.

Details

st_read_db 	 now a synonym for st_read

st_write_db 	 now a synonym for st_write

[Package sf version 0.9-3 Index]