sf-deprecated {sf}	R Documentation

Deprecated functions in sf

Description

These functions are provided for compatibility with older version of sf. They may eventually be completely removed.

Usage

st_read_db(conn = NULL, table = NULL, query = NULL, geom_column = NULL,
  EWKB = TRUE, ...)

Arguments

conn
open database connection

table
table name

query
SQL query to select records; see details

geom_column
character or integer: indicator of name or position of the geometry column; if not provided, the last column of type character is chosen

EWKB
logical; is the WKB is of type EWKB? if missing, defaults to TRUE

...
parameter(s) passed on to st_as_sf

Details

st_read_db 	 now a synonym for st_read

st_write_db 	 now a synonym for st_write

[Package sf version 0.6-2 Index]