st_crop {sf}	R Documentation

crop an sf object to a specific rectangle

Description

crop an sf object to a specific rectangle

Usage

st_crop(x, y, ..., xmin, ymin, xmax, ymax)

Arguments

x
object of class sf or sfc

y
numeric vector with named elements xmin, ymin, xmax and ymax, or object of class bbox

...
ignored

xmin
minimum x extent of cropping area

ymin
minimum y extent of cropping area

xmax
maximum x extent of cropping area

ymax
maximum y extent of cropping area

Details

setting arguments xmin, ymin, xmax and ymax implies that argument y gets ignored.

Examples

box = c(xmin = 0, ymin = 0, xmax = 1, ymax = 1)
pol = st_sfc(st_buffer(st_point(c(.5, .5)), .6))
pol_sf = st_sf(a=1, geom=pol)
plot(st_crop(pol, box))
plot(st_crop(pol_sf, st_bbox(box)))
# alternative:
plot(st_crop(pol, xmin = 0, ymin = 0, xmax = 1, ymax = 1))

[Package sf version 0.6-2 Index]