st_geometry_type {sf}	R Documentation

Return geometry type of an object

Description

Return geometry type of an object, as a factor

Usage

st_geometry_type(x, by_geometry = TRUE)

Arguments

x
object of class sf or sfc

by_geometry
logical; if TRUE, return geometry type of each geometry, else return geometry type of the set

Value

a factor with the geometry type of each simple feature geometry in x, or that of the whole set

[Package sf version 1.1-0 Index]