st_is_longlat {sf}	R Documentation

Assert whether simple feature coordinates are longlat degrees

Description

Assert whether simple feature coordinates are longlat degrees

Usage

st_is_longlat(x)

Arguments

x
object of class sf or sfc, or otherwise an object of a class that has an st_crs method returning a crs object

Value

TRUE if x has geographic coordinates, FALSE if it has projected coordinates, or NA if is.na(st_crs(x)).

[Package sf version 1.1-0 Index]