st_make_grid {sf}	R Documentation

Make a rectangular grid over the bounding box of a sf or sfc object

Description

Make a rectangular grid over the bounding box of a sf or sfc object

Usage

st_make_grid(x, cellsize = c(diff(st_bbox(x)[c(1, 3)]), diff(st_bbox(x)[c(2,
  4)]))/n, offset = st_bbox(x)[1:2], n = c(10, 10), crs = if (missing(x))
  NA_crs_ else st_crs(x), what = "polygons")

Arguments

x
object of class sf or sfc

cellsize
target cellsize

offset
numeric of lengt 2; lower left corner coordinates (x, y) of the grid

n
integer of length 1 or 2, number of grid cells in x and y direction (columns, rows)

crs
object of class crs; coordinate reference system of the target of the target grid in case argument x is missing, if x is not missing, its crs is inherited.

what
character; one of: "polygons", "corners", or "centers"

Value

Object of class sfc (simple feature geometry list column) with, depending on what, rectangular polygons, corner points of these polygons, or center points of these polygons.

Examples

plot(st_make_grid(what = "centers"), axes = TRUE)
plot(st_make_grid(what = "corners"), add = TRUE, col = 'green', pch=3)

[Package sf version 0.6-2 Index]