Return a collection of geometries resulting by splitting a geometry

Return a collection of geometries resulting by splitting a geometry

st_split(x, y)

Arguments

x
object with geometries to be splitted

y
object split with (blade); if y contains more than one feature geometry, the geometries are st_combined

Value

object of the same class as x

Details

st_split is only available if the package was linked against liblwgeom, which is currently not the case for the binary CRAN distributions; see the package source code repository for instructions how to install liblwgeom. The example below shows how to run-time check the availability of liblwgeom.

Examples

l = st_as_sfc('MULTILINESTRING((10 10, 190 190), (15 15, 30 30, 100 90))')
pt = st_sfc(st_point(c(30,30)))
if (!is.na(sf_extSoftVersion()["lwgeom"])) {
  st_split(l, pt)
}

#> Geometry set for 1 feature 
#> geometry type:  GEOMETRYCOLLECTION
#> dimension:      XY
#> bbox:           xmin: 10 ymin: 10 xmax: 190 ymax: 190
#> epsg (SRID):    NA
#> proj4string:    NA

#> GEOMETRYCOLLECTION(LINESTRING(30 30, 190 190), ...