st_transform {sf}	R Documentation

Transform or convert coordinates of simple feature

Description

Transform or convert coordinates of simple feature

Usage

st_transform(x, crs, ...)

## S3 method for class 'sfc'
st_transform(x, crs, ..., partial = TRUE, check = FALSE,
  use_gdal = TRUE)

## S3 method for class 'sf'
st_transform(x, crs, ...)

## S3 method for class 'sfg'
st_transform(x, crs, ...)

st_proj_info(type = "proj")

st_wrap_dateline(x, options, quiet)

## S3 method for class 'sfc'
st_wrap_dateline(x, options = "WRAPDATELINE=YES",
  quiet = TRUE)

## S3 method for class 'sf'
st_wrap_dateline(x, options = "WRAPDATELINE=YES", quiet = TRUE)

## S3 method for class 'sfg'
st_wrap_dateline(x, options = "WRAPDATELINE=YES",
  quiet = TRUE)

Arguments

x
object of class sf, sfc or sfg

crs
coordinate reference system: integer with the EPSG code, or character with proj4string

...
ignored

partial
logical; allow for partial projection, if not all points of a geometry can be projected (corresponds to setting environment variable OGR_ENABLE_PARTIAL_REPROJECTION to TRUE)

check
logical; perform a sanity check on resulting polygons?

use_gdal
logical; this parameter is deprecated. For transformations using PROJ.4 directly rather than indirectly through GDAL, use st_transform_proj of package lwgeom (see Details)

type
character; one of have_datum_files, proj, ellps, datum or units

options
character; should have "WRAPDATELINE=YES" to function; another parameter that is used is "DATELINEOFFSET=10" (where 10 is the default value)

quiet
logical; print options after they have been parsed?

Details

Transforms coordinates of object to new projection. Features that cannot be transformed are returned as empty geometries.

st_transform uses GDAL for coordinate transformations; internally, GDAL converts the proj4string into a well-known-text representation, before passing that on to PROJ.4. In this process, some information can get lost. Adding parameter +wktext to the proj4string definition may resolve this; see https://github.com/edzer/sp/issues/42.

Some PROJ.4 projections are not supported by GDAL, e.g. "+proj=wintri" because it does not have an inverse projection. Projecting to unsupported projections can be done by st_transform_proj, part of package lwgeom. Note that the unsupported proj4string cannot be passed as argument to st_crs, but has to be given as character string.

The st_transform method for sfg objects assumes that the CRS of the object is available as an attribute of that name.

st_proj_info lists the available projections, ellipses, datums or units supported by the Proj.4 library when type is equal to proj, ellps, datum or units; when type equals have_datum_files a boolean is returned indicating whether datum files are installed and accessible (checking for conus).

For a discussion of using options, see https://github.com/r-spatial/sf/issues/280 and https://github.com/r-spatial/sf/issues/541

Examples

p1 = st_point(c(7,52))
p2 = st_point(c(-30,20))
sfc = st_sfc(p1, p2, crs = 4326)
sfc
st_transform(sfc, 3857)
st_transform(st_sf(a=2:1, geom=sfc), "+init=epsg:3857")
nc = st_read(system.file("shape/nc.shp", package="sf"))
st_area(nc[1,]) # area from long/lat
st_area(st_transform(nc[1,], 32119)) # NC state plane, m
st_area(st_transform(nc[1,], 2264)) # NC state plane, US foot
library(units)
set_units(st_area(st_transform(nc[1,], 2264)), m^2)
st_transform(structure(p1, proj4string = "+init=epsg:4326"), "+init=epsg:3857")
st_proj_info("datum")
st_wrap_dateline(st_sfc(st_linestring(rbind(c(-179,0),c(179,0))), crs = 4326))
library(maps)
wrld <- st_as_sf(maps::map("world", fill = TRUE, plot = FALSE))
wrld_wrap <- st_wrap_dateline(wrld, options = c("WRAPDATELINE=YES", "DATELINEOFFSET=180"),
   quiet = TRUE)
wrld_moll <- st_transform(wrld_wrap, "+proj=moll")
plot(st_geometry(wrld_moll), col = "transparent")

[Package sf version 0.6-2 Index]