Summarize simple feature column

Summarize simple feature column

# S3 method for sfc
summary(object, ..., maxsum = 7L, maxp4s = 10L)

Arguments

object
object of class sfc

...
ignored

maxsum
maximum number of classes to summarize the simple feature column to

maxp4s
maximum number of characters to print from the PROJ.4 string