Methods to coerce simple feature geometries to corresponding Spatial* objects

Methods to coerce simple feature geometries to corresponding Spatial* objects

as_Spatial(from, cast = TRUE, IDs = paste0("ID", 1:length(from)))

Arguments

from
object of class sfc_POINT, sfc_MULTIPOINT, sfc_LINESTRING, sfc_MULTILINESTRING, sfc_POLYGON, or sfc_MULTIPOLYGON.

cast
logical; if TRUE, st_cast from before converting, so that e.g. GEOMETRY objects with a mix of POLYGON and MULTIPOLYGON are cast to MULTIPOLYGON.

IDs
character vector with IDs for the Spatial* geometries

Value

geometry-only object deriving from Spatial, of the appropriate class

Examples

nc = st_read(system.file("shape/nc.shp", package="sf"))

#> Reading layer `nc' from data source `/home/edzer/git/sf/inst/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> epsg (SRID):    4267
#> proj4string:    +proj=longlat +datum=NAD27 +no_defs

as_Spatial(st_geometry(nc[1,]))

#> An object of class "SpatialPolygons"
#> Slot "polygons":
#> [[1]]
#> An object of class "Polygons"
#> Slot "Polygons":
#> [[1]]
#> An object of class "Polygon"
#> Slot "labpt":
#> [1] -81.49826  36.43140
#> 
#> Slot "area":
#> [1] 0.1142835
#> 
#> Slot "hole":
#> [1] FALSE
#> 
#> Slot "ringDir":
#> [1] 1
#> 
#> Slot "coords":
#>            [,1]     [,2]
#>  [1,] -81.47276 36.23436
#>  [2,] -81.54084 36.27251
#>  [3,] -81.56198 36.27359
#>  [4,] -81.63306 36.34069
#>  [5,] -81.74107 36.39178
#>  [6,] -81.69828 36.47178
#>  [7,] -81.70280 36.51934
#>  [8,] -81.67000 36.58965
#>  [9,] -81.34530 36.57286
#> [10,] -81.34754 36.53791
#> [11,] -81.32478 36.51368
#> [12,] -81.31332 36.48070
#> [13,] -81.26624 36.43721
#> [14,] -81.26284 36.40504
#> [15,] -81.24069 36.37942
#> [16,] -81.23989 36.36536
#> [17,] -81.26424 36.35241
#> [18,] -81.32899 36.36350
#> [19,] -81.36137 36.35316
#> [20,] -81.36569 36.33905
#> [21,] -81.35413 36.29972
#> [22,] -81.36745 36.27870
#> [23,] -81.40639 36.28505
#> [24,] -81.41233 36.26729
#> [25,] -81.43104 36.26072
#> [26,] -81.45289 36.23959
#> [27,] -81.47276 36.23436
#> 
#> 
#> 
#> Slot "plotOrder":
#> [1] 1
#> 
#> Slot "labpt":
#> [1] -81.49826  36.43140
#> 
#> Slot "ID":
#> [1] "ID1"
#> 
#> Slot "area":
#> [1] 0.1142835
#> 
#> 
#> 
#> Slot "plotOrder":
#> [1] 1
#> 
#> Slot "bbox":
#>         min       max
#> x -81.74107 -81.23989
#> y  36.23436  36.58965
#> 
#> Slot "proj4string":
#> CRS arguments:
#>  +proj=longlat +datum=NAD27 +no_defs +ellps=clrk66
#> +nadgrids=@conus,@alaska,@ntv2_0.gsb,@ntv1_can.dat 
#>