Native interface to gdal utils

Native interface to gdal utils

gdal_utils(
  util = "info",
  source,
  destination,
  options = character(0),
  quiet = FALSE,
  processing = character(0),
  colorfilename = character(0)
)

Arguments

util
character; one of info, warp, rasterize, translate, vectortranslate, buildvrt, demprocessing, nearblack, grid

source
character; name of input layer(s); for warp or buidvrt this can be more than one

destination
character; name of output layer

options
character; raster layer read options

quiet
logical; if TRUE, suppress printing of output for info

processing
character; processing options for demprocessing

colorfilename
character; name of color file for demprocessing (mandatory if processing="color-relief")

Value

info returns a character vector with the raster metadata; all other utils return (invisibly) a logical indicating success (i.e., TRUE); in case of failure, an error is raised.