Dimension, simplicity or validity queries on simple feature geometries

Dimension, simplicity or validity queries on simple feature geometries

st_dimension(x, NA_if_empty = TRUE)

st_is_simple(x)

st_is_valid(x, NA_on_exception = TRUE, reason = FALSE)

Arguments

x
object of class sf, sfc or sfg

NA_if_empty
logical; if TRUE, return NA for empty geometries

NA_on_exception
logical; if TRUE, for polygons that would otherwise raise a GEOS error (exception, e.g. for a POLYGON having more than zero but less than 4 points, or a LINESTRING having one point) return an NA rather than raising an error, and suppress warning messages (e.g. about self-intersection); if FALSE, regular GEOS errors and warnings will be emitted.

reason
logical; if TRUE, return a character with, for each geometry, the reason for invalidity, NA on exception, or "Valid Geometry" otherwise.

Value

st_dimension returns a numeric vector with 0 for points, 1 for lines, 2 for surfaces, and, if NA_if_empty is TRUE, NA for empty geometries.

st_is_simple returns a logical vector, indicating for each geometry whether it is simple (e.g., not self-intersecting) st_is_vlaid returns a logical vector indicating for each geometries of x whether it is valid.

Examples

x = st_sfc(
        st_point(0:1),
        st_linestring(rbind(c(0,0),c(1,1))),
        st_polygon(list(rbind(c(0,0),c(1,0),c(0,1),c(0,0)))),
        st_multipoint(),
        st_linestring(),
        st_geometrycollection())
st_dimension(x)

#> [1]  0  1  2 NA NA NA

st_dimension(x, FALSE)

#> [1] 0 1 2 0 1 0

ls = st_linestring(rbind(c(0,0), c(1,1), c(1,0), c(0,1)))
st_is_simple(st_sfc(ls, st_point(c(0,0))))

#> [1] FALSE  TRUE

p1 = st_as_sfc("POLYGON((0 0, 0 10, 10 0, 10 10, 0 0))")
st_is_valid(p1)

#> Warning: Self-intersection at or near point 5 5

#> [1] FALSE

st_is_valid(st_sfc(st_point(0:1), p1[[1]]), reason = TRUE)

#> Warning: Self-intersection at or near point 5 5

#> [1] "Valid Geometry"         "Self-intersection[5 5]"