create sf object

create sf, which extends data.frame-like objects with a simple feature list column

st_sf(..., relation_to_geometry = NA_character_, row.names,
  stringsAsFactors = default.stringsAsFactors(), crs, precision)

# S3 method for sf
[(x, i, j, ..., drop)

Arguments

...

column elements to be binded into an sf object, one of them being of class sfc

relation_to_geometry

character vector; see details below.

row.names

row.names for the created sf object

stringsAsFactors

logical; see data.frame

crs

coordinate reference system: integer with the epsg code, or character with proj4string

precision

numeric; see st_as_binary

x

object of class sf

i

record selection, see [.data.frame

j

variable selection, see [.data.frame

drop

whether to drop to simpler (e.g. vector) representation, see [.data.frame

Details

relation_to_geometry specified for each non-geometry column how it relates to the geometry, and can have one of following values: "field", "lattice", "entity". "field" is used for attributes that are constant throughout the geometry (e.g. land use), "lattice" where the attribute is an aggregate value over the geometry (e.g. population density), "entity" when the attributes identifies the geometry of particular "thing", such as a building or a city. The default value, NA_character_, implies we don't know.

"[.sf" will return a data.frame if the geometry column (of class sfc) is dropped, an sfc object if only the geometry column is selected, otherwise the behavior depending on drop is identical to that of [.data.frame.

Examples

g = st_sfc(st_point(1:2))
st_sf(a=3,g)

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1 ymin: 2 xmax: 1 ymax: 2
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a          g
#> 1 3 POINT(1 2)
#> 

st_sf(g, a=3)

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1 ymin: 2 xmax: 1 ymax: 2
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a          g
#> 1 3 POINT(1 2)
#> 

st_sf(a=3, st_sfc(st_point(1:2))) # better to name it!

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1 ymin: 2 xmax: 1 ymax: 2
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a st_sfc.st_point.1.2..
#> 1 3            POINT(1 2)
#> 

g = st_sfc(st_point(1:2), st_point(3:4))
s = st_sf(a=3:4, g)
s[1,]

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1 ymin: 2 xmax: 1 ymax: 2
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a          g
#> 1 3 POINT(1 2)
#> 

class(s[1,])

#> [1] "sf"         "data.frame"
#> 

s[,1]

#>   x
#> 1 3
#> 2 4
#> 

class(s[,1])

#> [1] "data.frame"
#> 

s[,2]

#> Geometry set for 2 features 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1 ymin: 2 xmax: 3 ymax: 4
#> epsg (SRID):    NA
#> proj4string:    NA
#> POINT(1 2) 
#> POINT(3 4) 
#> 

class(s[,2])

#> [1] "sfc_POINT" "sfc"      
#> 

g = st_sf(a=2:3, g)
pol = st_sfc(st_polygon(list(cbind(c(0,3,3,0,0),c(0,0,3,3,0)))))
h = st_sf(r = 5, pol)
g[h,]

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1 ymin: 2 xmax: 1 ymax: 2
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a          g
#> 1 2 POINT(1 2)
#> 

h[g,]

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 3 ymax: 3
#> epsg (SRID):    NA
#> proj4string:    NA
#>   r                            pol
#> 1 5 POLYGON((0 0, 3 0, 3 3, 0 3...
#>