sf_extSoftVersion {sf}	R Documentation

Provide the external dependencies versions of the libraries linked to sf

Description

Provide the external dependencies versions of the libraries linked to sf

Usage

sf_extSoftVersion()

[Package sf version 0.6-2 Index]