sf_project {sf}	R Documentation

directly transform a set of coordinates

Description

directly transform a set of coordinates

Usage

sf_add_proj_units()

sf_project(
  from,
  to,
  pts,
  keep = FALSE,
  warn = TRUE,
  authority_compliant = st_axis_order()
)

Arguments

from
character description of source CRS, or object of class crs

to
character description of target CRS, or object of class crs

pts
two-column numeric matrix, or object that can be coerced into a matrix

keep
logical value controlling the handling of unprojectable points. If 'keep' is 'TRUE', then such points will yield 'Inf' or '-Inf' in the return value; otherwise an error is reported and nothing is returned.

warn
logical; if TRUE, warn when non-finite values are generated

authority_compliant
logical; TRUE means handle axis order authority compliant (e.g. EPSG:4326 implying x=lat, y=lon), FALSE means use visualisation order (i.e. always x=lon, y=lat)

Details

sf_add_proj_units loads the PROJ units 'link', 'us_in', 'ind_yd', 'ind_ft', and 'ind_ch' into the udunits database, and returns TRUE invisibly on success.

Value

two-column numeric matrix with transformed/converted coordinates, returning invalid values as Inf

[Package sf version 0.9-6 Index]