Create simple feature geometry list column

Create simple feature geometry list column, set class, and add coordinate reference system and precision

st_sfc(..., crs = NA_crs_, precision = 0, check_ring_dir = FALSE)

Arguments

...
zero or more simple feature geometries (objects of class sfg), or a single list of such objects; NULL values will get replaced by empty geometries.

crs
coordinate reference system: integer with the EPSG code, or character with proj4string

precision
numeric; see st_as_binary

check_ring_dir
see st_read

Value

an object of class sfc, which is a classed list-column with simple feature geometries.

Details

A simple feature geometry list-column is a list of class c("stc_TYPE", "sfc") which most often contains objects of identical type; in case of a mix of types or an empty set, TYPE is set to the superclass GEOMETRY.

Examples

pt1 = st_point(c(0,1))
pt2 = st_point(c(1,1))
(sfc = st_sfc(pt1, pt2))

#> Geometry set for 2 features 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 1 xmax: 1 ymax: 1
#> epsg (SRID):    NA
#> proj4string:    NA

#> POINT (0 1)

#> POINT (1 1)

d = st_sf(data.frame(a=1:2, geom=sfc))