Create simple feature collection object of class sfc from list

Create simple feature list column, set class, and add coordinate reference system

st_sfc(..., crs = NA_integer_, precision = 0)

Arguments

...

one or more simple feature geometries

crs

coordinate reference system: integer with the epsg code, or character with proj4string

precision

numeric; see st_as_binary

Details

a simple feature collection object is a list of class c("stc_TYPE", "sfc") which contains objects of identical type. This function creates such an object from a list of simple feature objects (of class sfg), and coerces their type if necessary: collections of XX and MULTIXX are coerced to MULTIXX (with XX: POINT, LINESTRING or POLYGON), other sets are coerced to GEOMETRYCOLLECTION.

in case epsg is given but proj4string is not and packages sp and rgdal can be loaded, the proj4string is expanded using the PROJ.4 epsg database.

Examples

pt1 = st_point(c(0,1))
pt2 = st_point(c(1,1))
(sfc = st_sfc(pt1, pt2))

#> Geometry set for 2 features 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 1 xmax: 1 ymax: 1
#> epsg (SRID):    NA
#> proj4string:    NA
#> POINT(0 1) 
#> POINT(1 1) 
#> 

d = data.frame(a = 1:2)