sgbp {sf}	R Documentation

Methods for dealing with sparse geometry binary predicate lists

Description

Methods for dealing with sparse geometry binary predicate lists

Usage

## S3 method for class 'sgbp'
print(x, ..., n = 10, max_nb = 10)

## S3 method for class 'sgbp'
t(x)

## S3 method for class 'sgbp'
as.matrix(x, ...)

## S3 method for class 'sgbp'
dim(x)

Arguments

x
object of class sgbp

...
ignored

n
integer; maximum number of items to print

max_nb
integer; maximum number of neighbours to print for each item

Details

sgbp are sparse matrices, stored as a list with integer vectors holding the ordered TRUE indices of each row. This means that for a dense, m x n matrix Q and a list L, if Q[i,j] is TRUE then j is an element of L[[i]]. Reversed: when k is the value of L[[i]][j], then Q[i,k] is TRUE.

[Package sf version 0.6-2 Index]