Convert foreign geometry object to an sfc object

Convert foreign geometry object to an sfc object

# S3 method for WKB
st_as_sfc(x, ..., EWKB = FALSE, pureR = FALSE,
  crs = NA_crs_)

# S3 method for character
st_as_sfc(x, crs = NA_integer_, ...)

st_as_sfc(x, ...)

# S3 method for SpatialPoints
st_as_sfc(x, ...)

# S3 method for SpatialPixels
st_as_sfc(x, ...)

# S3 method for SpatialMultiPoints
st_as_sfc(x, ...)

# S3 method for SpatialLines
st_as_sfc(x, ..., forceMulti = FALSE)

# S3 method for SpatialPolygons
st_as_sfc(x, ..., forceMulti = FALSE)

Arguments

x

object to convert

...

further arguments

EWKB

logical; if TRUE, parse as EWKB (PostGIS: ST_AsEWKB), otherwise as ISO WKB (PostGIS: ST_AsBinary)

pureR

logical; if TRUE, use only R code, if FALSE, use compiled (C++) code; use TRUE when the endian-ness of the binary differs from the host machine (.Platform$endian).

crs

integer or character; coordinate reference system for the geometry, see st_crs

forceMulti

logical; if TRUE, force coercion into MULTIPOLYGON or MULTILINE objects, else autodetect

Details

when converting from WKB, the object x is either a character vector such as typically obtained from PostGIS (either with leading "0x" or without), or a list with raw vectors representing the features in binary (raw) form.

if x is a character vector, it should be a vector containing the well-known-text representations of a single geometry for each vector element

Examples

wkb = structure(list("01010000204071000000000000801A064100000000AC5C1441"), class = "WKB")
st_as_sfc(wkb, EWKB = TRUE)

#> Geometry set for 1 feature 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 181072 ymin: 333611 xmax: 181072 ymax: 333611
#> epsg (SRID):    28992
#> proj4string:    +proj=sterea +lat_0=52.15616055555555 +lon_0=5.38763888888889 +k=0.9999079 +x_0=155000 +y_0=463000 +ellps=bessel +towgs84=565.4171,50.3319,465.5524,-0.398957,0.343988,-1.87740,4.0725 +units=m +no_defs

#> POINT(181072 333611)

wkb = structure(list("0x01010000204071000000000000801A064100000000AC5C1441"), class = "WKB")
st_as_sfc(wkb, EWKB = TRUE)

#> Geometry set for 1 feature 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 181072 ymin: 333611 xmax: 181072 ymax: 333611
#> epsg (SRID):    28992
#> proj4string:    +proj=sterea +lat_0=52.15616055555555 +lon_0=5.38763888888889 +k=0.9999079 +x_0=155000 +y_0=463000 +ellps=bessel +towgs84=565.4171,50.3319,465.5524,-0.398957,0.343988,-1.87740,4.0725 +units=m +no_defs

#> POINT(181072 333611)