st_drivers {sf}	R Documentation

Get GDAL drivers

Description

Get a list of the available GDAL drivers

Usage

st_drivers(what = "vector")

Arguments

what
character: "vector" or "raster", anything else will return all drivers.

Details

The drivers available will depend on the installation of GDAL/OGR, and can vary; the st_drivers() function shows which are available, and which may be written (but all are assumed to be readable). Note that stray files in data source directories (such as *.dbf) may lead to spurious errors that accompanying *.shp are missing.

field vsi refers to the driver's capability to read/create datasets through the VSI*L API.

Value

a data.frame with driver metadata

Examples

st_drivers()

[Package sf version 0.6-2 Index]