Get GDAL drivers

Get a list of the available GDAL drivers

st_drivers(what = "vector")

Arguments

what

character: "vector" or "raster", anything else will return all drivers.

Value

a data.frame with driver metadata

Details

The drivers available will depend on the installation of GDAL/OGR, and can vary; the st_drivers() function shows which are available, and which may be written (but all are assumed to be readable). Note that stray files in data source directories (such as *.dbf) may lead to suprious errors that accompanying *.shp are missing.

Examples

st_drivers()

#>               name                                            long_name write
#> 32          PCIDSK                                 PCIDSK Database File  TRUE
#> 40          netCDF                           Network Common Data Format  TRUE
#> 49     JP2OpenJPEG           JPEG-2000 driver based on OpenJPEG library FALSE
#> 69             PDF                                       Geospatial PDF  TRUE
#> 130 ESRI Shapefile                                       ESRI Shapefile  TRUE
#> 131   MapInfo File                                         MapInfo File  TRUE
#> 132        UK .NTF                                              UK .NTF FALSE
#> 133       OGR_SDTS                                                 SDTS FALSE
#> 134            S57                                       IHO S-57 (ENC)  TRUE
#> 135            DGN                                     Microstation DGN  TRUE
#> 136        OGR_VRT                             VRT - Virtual Datasource FALSE
#> 137            REC                                        EPIInfo .REC  FALSE
#> 138         Memory                                               Memory  TRUE
#> 139            BNA                                            Atlas BNA  TRUE
#> 140            CSV                         Comma Separated Value (.csv)  TRUE
#> 141            NAS                                          NAS - ALKIS FALSE
#> 142            GML                      Geography Markup Language (GML)  TRUE
#> 143            GPX                                                  GPX  TRUE
#> 144         LIBKML                     Keyhole Markup Language (LIBKML)  TRUE
#> 145            KML                        Keyhole Markup Language (KML)  TRUE
#> 146        GeoJSON                                              GeoJSON  TRUE
#> 147     Interlis 1                                           Interlis 1  TRUE
#> 148     Interlis 2                                           Interlis 2  TRUE
#> 149        OGR_GMT                             GMT ASCII Vectors (.gmt)  TRUE
#> 150           GPKG                                           GeoPackage  TRUE
#> 151         SQLite                                  SQLite / Spatialite  TRUE
#> 152       OGR_DODS                                             OGR_DODS FALSE
#> 153           ODBC                                                 ODBC  TRUE
#> 154           WAsP                                     WAsP .map format  TRUE
#> 155           PGeo                            ESRI Personal GeoDatabase FALSE
#> 156   MSSQLSpatial                Microsoft SQL Server Spatial Database  TRUE
#> 157       OGR_OGDI                        OGDI Vectors (VPF, VMAP, DCW) FALSE
#> 158     PostgreSQL                                   PostgreSQL/PostGIS  TRUE
#> 159          MySQL                                                MySQL  TRUE
#> 160    OpenFileGDB                                         ESRI FileGDB FALSE
#> 161         XPlane                 X-Plane/Flightgear aeronautical data FALSE
#> 162            DXF                                          AutoCAD DXF  TRUE
#> 163     Geoconcept                                           Geoconcept  TRUE
#> 164         GeoRSS                                               GeoRSS  TRUE
#> 165  GPSTrackMaker                                        GPSTrackMaker  TRUE
#> 166            VFK                 Czech Cadastral Exchange Data Format FALSE
#> 167         PGDUMP                                  PostgreSQL SQL dump  TRUE
#> 168            OSM                            OpenStreetMap XML and PBF FALSE
#> 169       GPSBabel                                             GPSBabel  TRUE
#> 170            SUA      Tim Newport-Peace's Special Use Airspace Format FALSE
#> 171        OpenAir                                              OpenAir FALSE
#> 172        OGR_PDS                         Planetary Data Systems TABLE FALSE
#> 173            WFS                        OGC WFS (Web Feature Service) FALSE
#> 174            HTF                         Hydrographic Transfer Vector FALSE
#> 175     AeronavFAA                                          Aeronav FAA FALSE
#> 176       Geomedia                                        Geomedia .mdb FALSE
#> 177         EDIGEO                        French EDIGEO exchange format FALSE
#> 178            GFT                                 Google Fusion Tables  TRUE
#> 179            SVG                             Scalable Vector Graphics FALSE
#> 180        CouchDB                                   CouchDB / GeoCouch  TRUE
#> 181       Cloudant                                   Cloudant / CouchDB  TRUE
#> 182         Idrisi                                 Idrisi Vector (.vct) FALSE
#> 183         ARCGEN                                    Arc/Info Generate FALSE
#> 184       SEGUKOOA                                 SEG-P1 / UKOOA P1/90 FALSE
#> 185           SEGY                                                SEG-Y FALSE
#> 186            XLS                                      MS Excel format FALSE
#> 187            ODS Open Document/ LibreOffice / OpenOffice Spreadsheet   TRUE
#> 188           XLSX                       MS Office Open XML spreadsheet  TRUE
#> 189  ElasticSearch                                       Elastic Search  TRUE
#> 190           Walk                                                 Walk FALSE
#> 191        CartoDB                                              CartoDB  TRUE
#> 192     AmigoCloud                                           AmigoCloud  TRUE
#> 193            SXF                          Storage and eXchange Format FALSE
#> 194        Selafin                                              Selafin  TRUE
#> 195            JML                                         OpenJUMP JML  TRUE
#> 196       PLSCENES                               Planet Labs Scenes API FALSE
#> 197            CSW               OGC CSW (Catalog  Service for the Web) FALSE
#> 198            VDV                  VDV-451/VDV-452/INTREST Data Format  TRUE
#> 199          TIGER                               U.S. Census TIGER/Line  TRUE
#> 200         AVCBin                             Arc/Info Binary Coverage FALSE
#> 201         AVCE00                        Arc/Info E00 (ASCII) Coverage FALSE
#> 202           HTTP                                HTTP Fetching Wrapper FALSE
#>      copy is_raster is_vector
#> 32  FALSE      TRUE      TRUE
#> 40   TRUE      TRUE      TRUE
#> 49   TRUE      TRUE      TRUE
#> 69   TRUE      TRUE      TRUE
#> 130 FALSE     FALSE      TRUE
#> 131 FALSE     FALSE      TRUE
#> 132 FALSE     FALSE      TRUE
#> 133 FALSE     FALSE      TRUE
#> 134 FALSE     FALSE      TRUE
#> 135 FALSE     FALSE      TRUE
#> 136 FALSE     FALSE      TRUE
#> 137 FALSE     FALSE      TRUE
#> 138 FALSE     FALSE      TRUE
#> 139 FALSE     FALSE      TRUE
#> 140 FALSE     FALSE      TRUE
#> 141 FALSE     FALSE      TRUE
#> 142 FALSE     FALSE      TRUE
#> 143 FALSE     FALSE      TRUE
#> 144 FALSE     FALSE      TRUE
#> 145 FALSE     FALSE      TRUE
#> 146 FALSE     FALSE      TRUE
#> 147 FALSE     FALSE      TRUE
#> 148 FALSE     FALSE      TRUE
#> 149 FALSE     FALSE      TRUE
#> 150  TRUE      TRUE      TRUE
#> 151 FALSE     FALSE      TRUE
#> 152 FALSE     FALSE      TRUE
#> 153 FALSE     FALSE      TRUE
#> 154 FALSE     FALSE      TRUE
#> 155 FALSE     FALSE      TRUE
#> 156 FALSE     FALSE      TRUE
#> 157 FALSE     FALSE      TRUE
#> 158 FALSE     FALSE      TRUE
#> 159 FALSE     FALSE      TRUE
#> 160 FALSE     FALSE      TRUE
#> 161 FALSE     FALSE      TRUE
#> 162 FALSE     FALSE      TRUE
#> 163 FALSE     FALSE      TRUE
#> 164 FALSE     FALSE      TRUE
#> 165 FALSE     FALSE      TRUE
#> 166 FALSE     FALSE      TRUE
#> 167 FALSE     FALSE      TRUE
#> 168 FALSE     FALSE      TRUE
#> 169 FALSE     FALSE      TRUE
#> 170 FALSE     FALSE      TRUE
#> 171 FALSE     FALSE      TRUE
#> 172 FALSE     FALSE      TRUE
#> 173 FALSE     FALSE      TRUE
#> 174 FALSE     FALSE      TRUE
#> 175 FALSE     FALSE      TRUE
#> 176 FALSE     FALSE      TRUE
#> 177 FALSE     FALSE      TRUE
#> 178 FALSE     FALSE      TRUE
#> 179 FALSE     FALSE      TRUE
#> 180 FALSE     FALSE      TRUE
#> 181 FALSE     FALSE      TRUE
#> 182 FALSE     FALSE      TRUE
#> 183 FALSE     FALSE      TRUE
#> 184 FALSE     FALSE      TRUE
#> 185 FALSE     FALSE      TRUE
#> 186 FALSE     FALSE      TRUE
#> 187 FALSE     FALSE      TRUE
#> 188 FALSE     FALSE      TRUE
#> 189 FALSE     FALSE      TRUE
#> 190 FALSE     FALSE      TRUE
#> 191 FALSE     FALSE      TRUE
#> 192 FALSE     FALSE      TRUE
#> 193 FALSE     FALSE      TRUE
#> 194 FALSE     FALSE      TRUE
#> 195 FALSE     FALSE      TRUE
#> 196 FALSE      TRUE      TRUE
#> 197 FALSE     FALSE      TRUE
#> 198 FALSE     FALSE      TRUE
#> 199 FALSE     FALSE      TRUE
#> 200 FALSE     FALSE      TRUE
#> 201 FALSE     FALSE      TRUE
#> 202 FALSE      TRUE      TRUE
#>