get GDAL drivers

get a list of the available GDAL drivers

st_drivers(what = "vector")

Arguments

what

character: "vector" or "raster", anything else will return all drivers.

Value

a data.frame with driver metadata

Details

The drivers available will depend on the installation of GDAL/OGR, and can vary; the st_drivers() function shows which are available, and which may be written (but all are assumed to be readable). Note that stray files in data source directories (such as *.dbf) may lead to suprious errors that accompanying *.shp are missing.

Examples

st_drivers()

#>               name                                                   long_name
#> 32          PCIDSK                                        PCIDSK Database File
#> 40          netCDF                                  Network Common Data Format
#> 49     JP2OpenJPEG                  JPEG-2000 driver based on OpenJPEG library
#> 53        JPEG2000 JPEG-2000 part 1 (ISO/IEC 15444-1), based on Jasper library
#> 69             PDF                                              Geospatial PDF
#> 128 ESRI Shapefile                                              ESRI Shapefile
#> 129   MapInfo File                                                MapInfo File
#> 130        UK .NTF                                                     UK .NTF
#> 131       OGR_SDTS                                                        SDTS
#> 132            S57                                              IHO S-57 (ENC)
#> 133            DGN                                            Microstation DGN
#> 134        OGR_VRT                                    VRT - Virtual Datasource
#> 135            REC                                               EPIInfo .REC 
#> 136         Memory                                                      Memory
#> 137            BNA                                                   Atlas BNA
#> 138            CSV                                Comma Separated Value (.csv)
#> 139            NAS                                                 NAS - ALKIS
#> 140            GML                             Geography Markup Language (GML)
#> 141            GPX                                                         GPX
#> 142         LIBKML                            Keyhole Markup Language (LIBKML)
#> 143            KML                               Keyhole Markup Language (KML)
#> 144        GeoJSON                                                     GeoJSON
#> 145     Interlis 1                                                  Interlis 1
#> 146     Interlis 2                                                  Interlis 2
#> 147        OGR_GMT                                    GMT ASCII Vectors (.gmt)
#> 148           GPKG                                                  GeoPackage
#> 149         SQLite                                         SQLite / Spatialite
#> 150           ODBC                                                        ODBC
#> 151           WAsP                                            WAsP .map format
#> 152           PGeo                                   ESRI Personal GeoDatabase
#> 153   MSSQLSpatial                       Microsoft SQL Server Spatial Database
#> 154       OGR_OGDI                               OGDI Vectors (VPF, VMAP, DCW)
#> 155     PostgreSQL                                          PostgreSQL/PostGIS
#> 156    OpenFileGDB                                                ESRI FileGDB
#> 157         XPlane                        X-Plane/Flightgear aeronautical data
#> 158            DXF                                                 AutoCAD DXF
#> 159     Geoconcept                                                  Geoconcept
#> 160         GeoRSS                                                      GeoRSS
#> 161  GPSTrackMaker                                               GPSTrackMaker
#> 162            VFK                        Czech Cadastral Exchange Data Format
#> 163         PGDUMP                                         PostgreSQL SQL dump
#> 164            OSM                                   OpenStreetMap XML and PBF
#> 165       GPSBabel                                                    GPSBabel
#> 166            SUA             Tim Newport-Peace's Special Use Airspace Format
#> 167        OpenAir                                                     OpenAir
#> 168        OGR_PDS                                Planetary Data Systems TABLE
#> 169            WFS                               OGC WFS (Web Feature Service)
#> 170            HTF                                Hydrographic Transfer Vector
#> 171     AeronavFAA                                                 Aeronav FAA
#> 172       Geomedia                                               Geomedia .mdb
#> 173         EDIGEO                               French EDIGEO exchange format
#> 174            GFT                                        Google Fusion Tables
#> 175            SVG                                    Scalable Vector Graphics
#> 176        CouchDB                                          CouchDB / GeoCouch
#> 177       Cloudant                                          Cloudant / CouchDB
#> 178         Idrisi                                        Idrisi Vector (.vct)
#> 179         ARCGEN                                           Arc/Info Generate
#> 180       SEGUKOOA                                        SEG-P1 / UKOOA P1/90
#> 181           SEGY                                                       SEG-Y
#> 182            XLS                                             MS Excel format
#> 183            ODS        Open Document/ LibreOffice / OpenOffice Spreadsheet 
#> 184           XLSX                              MS Office Open XML spreadsheet
#> 185  ElasticSearch                                              Elastic Search
#> 186           Walk                                                        Walk
#> 187        CartoDB                                                     CartoDB
#> 188     AmigoCloud                                                  AmigoCloud
#> 189            SXF                                 Storage and eXchange Format
#> 190        Selafin                                                     Selafin
#> 191            JML                                                OpenJUMP JML
#> 192       PLSCENES                                      Planet Labs Scenes API
#> 193            CSW                      OGC CSW (Catalog  Service for the Web)
#> 194            VDV                         VDV-451/VDV-452/INTREST Data Format
#> 195          TIGER                                      U.S. Census TIGER/Line
#> 196         AVCBin                                    Arc/Info Binary Coverage
#> 197         AVCE00                               Arc/Info E00 (ASCII) Coverage
#> 198           HTTP                                       HTTP Fetching Wrapper
#>     write  copy is_raster is_vector
#> 32   TRUE FALSE      TRUE      TRUE
#> 40   TRUE  TRUE      TRUE      TRUE
#> 49  FALSE  TRUE      TRUE      TRUE
#> 53  FALSE  TRUE      TRUE      TRUE
#> 69   TRUE  TRUE      TRUE      TRUE
#> 128  TRUE FALSE     FALSE      TRUE
#> 129  TRUE FALSE     FALSE      TRUE
#> 130 FALSE FALSE     FALSE      TRUE
#> 131 FALSE FALSE     FALSE      TRUE
#> 132  TRUE FALSE     FALSE      TRUE
#> 133  TRUE FALSE     FALSE      TRUE
#> 134 FALSE FALSE     FALSE      TRUE
#> 135 FALSE FALSE     FALSE      TRUE
#> 136  TRUE FALSE     FALSE      TRUE
#> 137  TRUE FALSE     FALSE      TRUE
#> 138  TRUE FALSE     FALSE      TRUE
#> 139 FALSE FALSE     FALSE      TRUE
#> 140  TRUE FALSE     FALSE      TRUE
#> 141  TRUE FALSE     FALSE      TRUE
#> 142  TRUE FALSE     FALSE      TRUE
#> 143  TRUE FALSE     FALSE      TRUE
#> 144  TRUE FALSE     FALSE      TRUE
#> 145  TRUE FALSE     FALSE      TRUE
#> 146  TRUE FALSE     FALSE      TRUE
#> 147  TRUE FALSE     FALSE      TRUE
#> 148  TRUE  TRUE      TRUE      TRUE
#> 149  TRUE FALSE     FALSE      TRUE
#> 150  TRUE FALSE     FALSE      TRUE
#> 151  TRUE FALSE     FALSE      TRUE
#> 152 FALSE FALSE     FALSE      TRUE
#> 153  TRUE FALSE     FALSE      TRUE
#> 154 FALSE FALSE     FALSE      TRUE
#> 155  TRUE FALSE     FALSE      TRUE
#> 156 FALSE FALSE     FALSE      TRUE
#> 157 FALSE FALSE     FALSE      TRUE
#> 158  TRUE FALSE     FALSE      TRUE
#> 159  TRUE FALSE     FALSE      TRUE
#> 160  TRUE FALSE     FALSE      TRUE
#> 161  TRUE FALSE     FALSE      TRUE
#> 162 FALSE FALSE     FALSE      TRUE
#> 163  TRUE FALSE     FALSE      TRUE
#> 164 FALSE FALSE     FALSE      TRUE
#> 165  TRUE FALSE     FALSE      TRUE
#> 166 FALSE FALSE     FALSE      TRUE
#> 167 FALSE FALSE     FALSE      TRUE
#> 168 FALSE FALSE     FALSE      TRUE
#> 169 FALSE FALSE     FALSE      TRUE
#> 170 FALSE FALSE     FALSE      TRUE
#> 171 FALSE FALSE     FALSE      TRUE
#> 172 FALSE FALSE     FALSE      TRUE
#> 173 FALSE FALSE     FALSE      TRUE
#> 174  TRUE FALSE     FALSE      TRUE
#> 175 FALSE FALSE     FALSE      TRUE
#> 176  TRUE FALSE     FALSE      TRUE
#> 177  TRUE FALSE     FALSE      TRUE
#> 178 FALSE FALSE     FALSE      TRUE
#> 179 FALSE FALSE     FALSE      TRUE
#> 180 FALSE FALSE     FALSE      TRUE
#> 181 FALSE FALSE     FALSE      TRUE
#> 182 FALSE FALSE     FALSE      TRUE
#> 183  TRUE FALSE     FALSE      TRUE
#> 184  TRUE FALSE     FALSE      TRUE
#> 185  TRUE FALSE     FALSE      TRUE
#> 186 FALSE FALSE     FALSE      TRUE
#> 187  TRUE FALSE     FALSE      TRUE
#> 188  TRUE FALSE     FALSE      TRUE
#> 189 FALSE FALSE     FALSE      TRUE
#> 190  TRUE FALSE     FALSE      TRUE
#> 191  TRUE FALSE     FALSE      TRUE
#> 192 FALSE FALSE      TRUE      TRUE
#> 193 FALSE FALSE     FALSE      TRUE
#> 194  TRUE FALSE     FALSE      TRUE
#> 195  TRUE FALSE     FALSE      TRUE
#> 196 FALSE FALSE     FALSE      TRUE
#> 197 FALSE FALSE     FALSE      TRUE
#> 198 FALSE FALSE      TRUE      TRUE
#>