Get GDAL drivers

Get a list of the available GDAL drivers

st_drivers(what = "vector")

Arguments

what
character: "vector" or "raster", anything else will return all drivers.

Value

A data.frame with driver metadata.

Details

The drivers available will depend on the installation of GDAL/OGR, and can vary; the st_drivers() function shows all the drivers that are readable, and which may be written. The field vsi refers to the driver's capability to read/create datasets through the VSI*L API. See GDAL website for additional details on driver support.

Examples

st_drivers()

#>                          name
#> PCIDSK                 PCIDSK
#> netCDF                 netCDF
#> JP2OpenJPEG       JP2OpenJPEG
#> PDF                       PDF
#> ESRI Shapefile ESRI Shapefile
#> MapInfo File     MapInfo File
#> UK .NTF               UK .NTF
#> OGR_SDTS             OGR_SDTS
#> S57                       S57
#> DGN                       DGN
#> OGR_VRT               OGR_VRT
#> REC                       REC
#> Memory                 Memory
#> BNA                       BNA
#> CSV                       CSV
#> NAS                       NAS
#> GML                       GML
#> GPX                       GPX
#> LIBKML                 LIBKML
#> KML                       KML
#> GeoJSON               GeoJSON
#> Interlis 1         Interlis 1
#> Interlis 2         Interlis 2
#> OGR_GMT               OGR_GMT
#> GPKG                     GPKG
#> SQLite                 SQLite
#> OGR_DODS             OGR_DODS
#> ODBC                     ODBC
#> WAsP                     WAsP
#> PGeo                     PGeo
#> MSSQLSpatial     MSSQLSpatial
#> OGR_OGDI             OGR_OGDI
#> PostgreSQL         PostgreSQL
#> MySQL                   MySQL
#> OpenFileGDB       OpenFileGDB
#> XPlane                 XPlane
#> DXF                       DXF
#> CAD                       CAD
#> Geoconcept         Geoconcept
#> GeoRSS                 GeoRSS
#> GPSTrackMaker   GPSTrackMaker
#> VFK                       VFK
#> PGDUMP                 PGDUMP
#> OSM                       OSM
#> GPSBabel             GPSBabel
#> SUA                       SUA
#> OpenAir               OpenAir
#> OGR_PDS               OGR_PDS
#> WFS                       WFS
#> HTF                       HTF
#> AeronavFAA         AeronavFAA
#> Geomedia             Geomedia
#> EDIGEO                 EDIGEO
#> GFT                       GFT
#> SVG                       SVG
#> CouchDB               CouchDB
#> Cloudant             Cloudant
#> Idrisi                 Idrisi
#> ARCGEN                 ARCGEN
#> SEGUKOOA             SEGUKOOA
#> SEGY                     SEGY
#> XLS                       XLS
#> ODS                       ODS
#> XLSX                     XLSX
#> ElasticSearch   ElasticSearch
#> Walk                     Walk
#> Carto                   Carto
#> AmigoCloud         AmigoCloud
#> SXF                       SXF
#> Selafin               Selafin
#> JML                       JML
#> PLSCENES             PLSCENES
#> CSW                       CSW
#> VDV                       VDV
#> GMLAS                   GMLAS
#> TIGER                   TIGER
#> AVCBin                 AVCBin
#> AVCE00                 AVCE00
#> HTTP                     HTTP
#>                                                                    long_name
#> PCIDSK                                                  PCIDSK Database File
#> netCDF                                            Network Common Data Format
#> JP2OpenJPEG                       JPEG-2000 driver based on OpenJPEG library
#> PDF                                                           Geospatial PDF
#> ESRI Shapefile                                                ESRI Shapefile
#> MapInfo File                                                    MapInfo File
#> UK .NTF                                                              UK .NTF
#> OGR_SDTS                                                                SDTS
#> S57                                                           IHO S-57 (ENC)
#> DGN                                                         Microstation DGN
#> OGR_VRT                                             VRT - Virtual Datasource
#> REC                                                            EPIInfo .REC 
#> Memory                                                                Memory
#> BNA                                                                Atlas BNA
#> CSV                                             Comma Separated Value (.csv)
#> NAS                                                              NAS - ALKIS
#> GML                                          Geography Markup Language (GML)
#> GPX                                                                      GPX
#> LIBKML                                      Keyhole Markup Language (LIBKML)
#> KML                                            Keyhole Markup Language (KML)
#> GeoJSON                                                              GeoJSON
#> Interlis 1                                                        Interlis 1
#> Interlis 2                                                        Interlis 2
#> OGR_GMT                                             GMT ASCII Vectors (.gmt)
#> GPKG                                                              GeoPackage
#> SQLite                                                   SQLite / Spatialite
#> OGR_DODS                                                            OGR_DODS
#> ODBC                                                                    ODBC
#> WAsP                                                        WAsP .map format
#> PGeo                                               ESRI Personal GeoDatabase
#> MSSQLSpatial                           Microsoft SQL Server Spatial Database
#> OGR_OGDI                                       OGDI Vectors (VPF, VMAP, DCW)
#> PostgreSQL                                                PostgreSQL/PostGIS
#> MySQL                                                                  MySQL
#> OpenFileGDB                                                     ESRI FileGDB
#> XPlane                                  X-Plane/Flightgear aeronautical data
#> DXF                                                              AutoCAD DXF
#> CAD                                                           AutoCAD Driver
#> Geoconcept                                                        Geoconcept
#> GeoRSS                                                                GeoRSS
#> GPSTrackMaker                                                  GPSTrackMaker
#> VFK                                     Czech Cadastral Exchange Data Format
#> PGDUMP                                                   PostgreSQL SQL dump
#> OSM                                                OpenStreetMap XML and PBF
#> GPSBabel                                                            GPSBabel
#> SUA                          Tim Newport-Peace's Special Use Airspace Format
#> OpenAir                                                              OpenAir
#> OGR_PDS                                         Planetary Data Systems TABLE
#> WFS                                            OGC WFS (Web Feature Service)
#> HTF                                             Hydrographic Transfer Vector
#> AeronavFAA                                                       Aeronav FAA
#> Geomedia                                                       Geomedia .mdb
#> EDIGEO                                         French EDIGEO exchange format
#> GFT                                                     Google Fusion Tables
#> SVG                                                 Scalable Vector Graphics
#> CouchDB                                                   CouchDB / GeoCouch
#> Cloudant                                                  Cloudant / CouchDB
#> Idrisi                                                  Idrisi Vector (.vct)
#> ARCGEN                                                     Arc/Info Generate
#> SEGUKOOA                                                SEG-P1 / UKOOA P1/90
#> SEGY                                                                   SEG-Y
#> XLS                                                          MS Excel format
#> ODS                     Open Document/ LibreOffice / OpenOffice Spreadsheet 
#> XLSX                                          MS Office Open XML spreadsheet
#> ElasticSearch                                                 Elastic Search
#> Walk                                                                    Walk
#> Carto                                                                  Carto
#> AmigoCloud                                                        AmigoCloud
#> SXF                                              Storage and eXchange Format
#> Selafin                                                              Selafin
#> JML                                                             OpenJUMP JML
#> PLSCENES                                              Planet Labs Scenes API
#> CSW                                   OGC CSW (Catalog  Service for the Web)
#> VDV                                      VDV-451/VDV-452/INTREST Data Format
#> GMLAS          Geography Markup Language (GML) driven by application schemas
#> TIGER                                                 U.S. Census TIGER/Line
#> AVCBin                                              Arc/Info Binary Coverage
#> AVCE00                                         Arc/Info E00 (ASCII) Coverage
#> HTTP                                                   HTTP Fetching Wrapper
#>                write  copy is_raster is_vector   vsi
#> PCIDSK          TRUE FALSE      TRUE      TRUE  TRUE
#> netCDF          TRUE  TRUE      TRUE      TRUE FALSE
#> JP2OpenJPEG    FALSE  TRUE      TRUE      TRUE  TRUE
#> PDF             TRUE  TRUE      TRUE      TRUE  TRUE
#> ESRI Shapefile  TRUE FALSE     FALSE      TRUE  TRUE
#> MapInfo File    TRUE FALSE     FALSE      TRUE  TRUE
#> UK .NTF        FALSE FALSE     FALSE      TRUE FALSE
#> OGR_SDTS       FALSE FALSE     FALSE      TRUE FALSE
#> S57             TRUE FALSE     FALSE      TRUE  TRUE
#> DGN             TRUE FALSE     FALSE      TRUE FALSE
#> OGR_VRT        FALSE FALSE     FALSE      TRUE  TRUE
#> REC            FALSE FALSE     FALSE      TRUE FALSE
#> Memory          TRUE FALSE     FALSE      TRUE FALSE
#> BNA             TRUE FALSE     FALSE      TRUE  TRUE
#> CSV             TRUE FALSE     FALSE      TRUE  TRUE
#> NAS            FALSE FALSE     FALSE      TRUE FALSE
#> GML             TRUE FALSE     FALSE      TRUE  TRUE
#> GPX             TRUE FALSE     FALSE      TRUE  TRUE
#> LIBKML          TRUE FALSE     FALSE      TRUE  TRUE
#> KML             TRUE FALSE     FALSE      TRUE  TRUE
#> GeoJSON         TRUE FALSE     FALSE      TRUE  TRUE
#> Interlis 1      TRUE FALSE     FALSE      TRUE FALSE
#> Interlis 2      TRUE FALSE     FALSE      TRUE FALSE
#> OGR_GMT         TRUE FALSE     FALSE      TRUE FALSE
#> GPKG            TRUE  TRUE      TRUE      TRUE  TRUE
#> SQLite          TRUE FALSE     FALSE      TRUE  TRUE
#> OGR_DODS       FALSE FALSE     FALSE      TRUE FALSE
#> ODBC            TRUE FALSE     FALSE      TRUE FALSE
#> WAsP            TRUE FALSE     FALSE      TRUE  TRUE
#> PGeo           FALSE FALSE     FALSE      TRUE FALSE
#> MSSQLSpatial    TRUE FALSE     FALSE      TRUE FALSE
#> OGR_OGDI       FALSE FALSE     FALSE      TRUE FALSE
#> PostgreSQL      TRUE FALSE     FALSE      TRUE FALSE
#> MySQL           TRUE FALSE     FALSE      TRUE FALSE
#> OpenFileGDB    FALSE FALSE     FALSE      TRUE  TRUE
#> XPlane         FALSE FALSE     FALSE      TRUE  TRUE
#> DXF             TRUE FALSE     FALSE      TRUE  TRUE
#> CAD            FALSE FALSE      TRUE      TRUE  TRUE
#> Geoconcept      TRUE FALSE     FALSE      TRUE FALSE
#> GeoRSS          TRUE FALSE     FALSE      TRUE  TRUE
#> GPSTrackMaker   TRUE FALSE     FALSE      TRUE  TRUE
#> VFK            FALSE FALSE     FALSE      TRUE FALSE
#> PGDUMP          TRUE FALSE     FALSE      TRUE  TRUE
#> OSM            FALSE FALSE     FALSE      TRUE  TRUE
#> GPSBabel        TRUE FALSE     FALSE      TRUE FALSE
#> SUA            FALSE FALSE     FALSE      TRUE  TRUE
#> OpenAir        FALSE FALSE     FALSE      TRUE  TRUE
#> OGR_PDS        FALSE FALSE     FALSE      TRUE  TRUE
#> WFS            FALSE FALSE     FALSE      TRUE  TRUE
#> HTF            FALSE FALSE     FALSE      TRUE  TRUE
#> AeronavFAA     FALSE FALSE     FALSE      TRUE  TRUE
#> Geomedia       FALSE FALSE     FALSE      TRUE FALSE
#> EDIGEO         FALSE FALSE     FALSE      TRUE  TRUE
#> GFT             TRUE FALSE     FALSE      TRUE FALSE
#> SVG            FALSE FALSE     FALSE      TRUE  TRUE
#> CouchDB         TRUE FALSE     FALSE      TRUE FALSE
#> Cloudant        TRUE FALSE     FALSE      TRUE FALSE
#> Idrisi         FALSE FALSE     FALSE      TRUE  TRUE
#> ARCGEN         FALSE FALSE     FALSE      TRUE  TRUE
#> SEGUKOOA       FALSE FALSE     FALSE      TRUE  TRUE
#> SEGY           FALSE FALSE     FALSE      TRUE  TRUE
#> XLS            FALSE FALSE     FALSE      TRUE FALSE
#> ODS             TRUE FALSE     FALSE      TRUE  TRUE
#> XLSX            TRUE FALSE     FALSE      TRUE  TRUE
#> ElasticSearch   TRUE FALSE     FALSE      TRUE FALSE
#> Walk           FALSE FALSE     FALSE      TRUE FALSE
#> Carto           TRUE FALSE     FALSE      TRUE FALSE
#> AmigoCloud      TRUE FALSE     FALSE      TRUE FALSE
#> SXF            FALSE FALSE     FALSE      TRUE FALSE
#> Selafin         TRUE FALSE     FALSE      TRUE  TRUE
#> JML             TRUE FALSE     FALSE      TRUE  TRUE
#> PLSCENES       FALSE FALSE      TRUE      TRUE FALSE
#> CSW            FALSE FALSE     FALSE      TRUE FALSE
#> VDV             TRUE FALSE     FALSE      TRUE  TRUE
#> GMLAS          FALSE  TRUE     FALSE      TRUE  TRUE
#> TIGER           TRUE FALSE     FALSE      TRUE  TRUE
#> AVCBin         FALSE FALSE     FALSE      TRUE FALSE
#> AVCE00         FALSE FALSE     FALSE      TRUE FALSE
#> HTTP           FALSE FALSE      TRUE      TRUE FALSE