Get GDAL drivers

Get a list of the available GDAL drivers

st_drivers(what = "vector")

Arguments

what

character: "vector" or "raster", anything else will return all drivers.

Value

a data.frame with driver metadata

Details

The drivers available will depend on the installation of GDAL/OGR, and can vary; the st_drivers() function shows which are available, and which may be written (but all are assumed to be readable). Note that stray files in data source directories (such as *.dbf) may lead to suprious errors that accompanying *.shp are missing.

Examples

st_drivers()

#>                          name
#> PCIDSK                 PCIDSK
#> netCDF                 netCDF
#> JP2OpenJPEG       JP2OpenJPEG
#> PDF                       PDF
#> ESRI Shapefile ESRI Shapefile
#> MapInfo File     MapInfo File
#> UK .NTF               UK .NTF
#> OGR_SDTS             OGR_SDTS
#> S57                       S57
#> DGN                       DGN
#> OGR_VRT               OGR_VRT
#> REC                       REC
#> Memory                 Memory
#> BNA                       BNA
#> CSV                       CSV
#> NAS                       NAS
#> GML                       GML
#> GPX                       GPX
#> LIBKML                 LIBKML
#> KML                       KML
#> GeoJSON               GeoJSON
#> Interlis 1         Interlis 1
#> Interlis 2         Interlis 2
#> OGR_GMT               OGR_GMT
#> GPKG                     GPKG
#> SQLite                 SQLite
#> ODBC                     ODBC
#> WAsP                     WAsP
#> PGeo                     PGeo
#> MSSQLSpatial     MSSQLSpatial
#> OGR_OGDI             OGR_OGDI
#> PostgreSQL         PostgreSQL
#> OpenFileGDB       OpenFileGDB
#> XPlane                 XPlane
#> DXF                       DXF
#> Geoconcept         Geoconcept
#> GeoRSS                 GeoRSS
#> GPSTrackMaker   GPSTrackMaker
#> VFK                       VFK
#> PGDUMP                 PGDUMP
#> OSM                       OSM
#> GPSBabel             GPSBabel
#> SUA                       SUA
#> OpenAir               OpenAir
#> OGR_PDS               OGR_PDS
#> WFS                       WFS
#> HTF                       HTF
#> AeronavFAA         AeronavFAA
#> Geomedia             Geomedia
#> EDIGEO                 EDIGEO
#> GFT                       GFT
#> SVG                       SVG
#> CouchDB               CouchDB
#> Cloudant             Cloudant
#> Idrisi                 Idrisi
#> ARCGEN                 ARCGEN
#> SEGUKOOA             SEGUKOOA
#> SEGY                     SEGY
#> XLS                       XLS
#> ODS                       ODS
#> XLSX                     XLSX
#> ElasticSearch   ElasticSearch
#> Walk                     Walk
#> Carto                   Carto
#> AmigoCloud         AmigoCloud
#> SXF                       SXF
#> Selafin               Selafin
#> JML                       JML
#> PLSCENES             PLSCENES
#> CSW                       CSW
#> VDV                       VDV
#> TIGER                   TIGER
#> AVCBin                 AVCBin
#> AVCE00                 AVCE00
#> HTTP                     HTTP
#>                                                           long_name write  copy
#> PCIDSK                                         PCIDSK Database File  TRUE FALSE
#> netCDF                                   Network Common Data Format  TRUE  TRUE
#> JP2OpenJPEG              JPEG-2000 driver based on OpenJPEG library FALSE  TRUE
#> PDF                                                  Geospatial PDF  TRUE  TRUE
#> ESRI Shapefile                                       ESRI Shapefile  TRUE FALSE
#> MapInfo File                                           MapInfo File  TRUE FALSE
#> UK .NTF                                                     UK .NTF FALSE FALSE
#> OGR_SDTS                                                       SDTS FALSE FALSE
#> S57                                                  IHO S-57 (ENC)  TRUE FALSE
#> DGN                                                Microstation DGN  TRUE FALSE
#> OGR_VRT                                    VRT - Virtual Datasource FALSE FALSE
#> REC                                                   EPIInfo .REC  FALSE FALSE
#> Memory                                                       Memory  TRUE FALSE
#> BNA                                                       Atlas BNA  TRUE FALSE
#> CSV                                    Comma Separated Value (.csv)  TRUE FALSE
#> NAS                                                     NAS - ALKIS FALSE FALSE
#> GML                                 Geography Markup Language (GML)  TRUE FALSE
#> GPX                                                             GPX  TRUE FALSE
#> LIBKML                             Keyhole Markup Language (LIBKML)  TRUE FALSE
#> KML                                   Keyhole Markup Language (KML)  TRUE FALSE
#> GeoJSON                                                     GeoJSON  TRUE FALSE
#> Interlis 1                                               Interlis 1  TRUE FALSE
#> Interlis 2                                               Interlis 2  TRUE FALSE
#> OGR_GMT                                    GMT ASCII Vectors (.gmt)  TRUE FALSE
#> GPKG                                                     GeoPackage  TRUE  TRUE
#> SQLite                                          SQLite / Spatialite  TRUE FALSE
#> ODBC                                                           ODBC  TRUE FALSE
#> WAsP                                               WAsP .map format  TRUE FALSE
#> PGeo                                      ESRI Personal GeoDatabase FALSE FALSE
#> MSSQLSpatial                  Microsoft SQL Server Spatial Database  TRUE FALSE
#> OGR_OGDI                              OGDI Vectors (VPF, VMAP, DCW) FALSE FALSE
#> PostgreSQL                                       PostgreSQL/PostGIS  TRUE FALSE
#> OpenFileGDB                                            ESRI FileGDB FALSE FALSE
#> XPlane                         X-Plane/Flightgear aeronautical data FALSE FALSE
#> DXF                                                     AutoCAD DXF  TRUE FALSE
#> Geoconcept                                               Geoconcept  TRUE FALSE
#> GeoRSS                                                       GeoRSS  TRUE FALSE
#> GPSTrackMaker                                         GPSTrackMaker  TRUE FALSE
#> VFK                            Czech Cadastral Exchange Data Format FALSE FALSE
#> PGDUMP                                          PostgreSQL SQL dump  TRUE FALSE
#> OSM                                       OpenStreetMap XML and PBF FALSE FALSE
#> GPSBabel                                                   GPSBabel  TRUE FALSE
#> SUA                 Tim Newport-Peace's Special Use Airspace Format FALSE FALSE
#> OpenAir                                                     OpenAir FALSE FALSE
#> OGR_PDS                                Planetary Data Systems TABLE FALSE FALSE
#> WFS                                   OGC WFS (Web Feature Service) FALSE FALSE
#> HTF                                    Hydrographic Transfer Vector FALSE FALSE
#> AeronavFAA                                              Aeronav FAA FALSE FALSE
#> Geomedia                                              Geomedia .mdb FALSE FALSE
#> EDIGEO                                French EDIGEO exchange format FALSE FALSE
#> GFT                                            Google Fusion Tables  TRUE FALSE
#> SVG                                        Scalable Vector Graphics FALSE FALSE
#> CouchDB                                          CouchDB / GeoCouch  TRUE FALSE
#> Cloudant                                         Cloudant / CouchDB  TRUE FALSE
#> Idrisi                                         Idrisi Vector (.vct) FALSE FALSE
#> ARCGEN                                            Arc/Info Generate FALSE FALSE
#> SEGUKOOA                                       SEG-P1 / UKOOA P1/90 FALSE FALSE
#> SEGY                                                          SEG-Y FALSE FALSE
#> XLS                                                 MS Excel format FALSE FALSE
#> ODS            Open Document/ LibreOffice / OpenOffice Spreadsheet   TRUE FALSE
#> XLSX                                 MS Office Open XML spreadsheet  TRUE FALSE
#> ElasticSearch                                        Elastic Search  TRUE FALSE
#> Walk                                                           Walk FALSE FALSE
#> Carto                                                         Carto  TRUE FALSE
#> AmigoCloud                                               AmigoCloud  TRUE FALSE
#> SXF                                     Storage and eXchange Format FALSE FALSE
#> Selafin                                                     Selafin  TRUE FALSE
#> JML                                                    OpenJUMP JML  TRUE FALSE
#> PLSCENES                                     Planet Labs Scenes API FALSE FALSE
#> CSW                          OGC CSW (Catalog  Service for the Web) FALSE FALSE
#> VDV                             VDV-451/VDV-452/INTREST Data Format  TRUE FALSE
#> TIGER                                        U.S. Census TIGER/Line  TRUE FALSE
#> AVCBin                                     Arc/Info Binary Coverage FALSE FALSE
#> AVCE00                                Arc/Info E00 (ASCII) Coverage FALSE FALSE
#> HTTP                                          HTTP Fetching Wrapper FALSE FALSE
#>                is_raster is_vector
#> PCIDSK              TRUE      TRUE
#> netCDF              TRUE      TRUE
#> JP2OpenJPEG         TRUE      TRUE
#> PDF                 TRUE      TRUE
#> ESRI Shapefile     FALSE      TRUE
#> MapInfo File       FALSE      TRUE
#> UK .NTF            FALSE      TRUE
#> OGR_SDTS           FALSE      TRUE
#> S57                FALSE      TRUE
#> DGN                FALSE      TRUE
#> OGR_VRT            FALSE      TRUE
#> REC                FALSE      TRUE
#> Memory             FALSE      TRUE
#> BNA                FALSE      TRUE
#> CSV                FALSE      TRUE
#> NAS                FALSE      TRUE
#> GML                FALSE      TRUE
#> GPX                FALSE      TRUE
#> LIBKML             FALSE      TRUE
#> KML                FALSE      TRUE
#> GeoJSON            FALSE      TRUE
#> Interlis 1         FALSE      TRUE
#> Interlis 2         FALSE      TRUE
#> OGR_GMT            FALSE      TRUE
#> GPKG                TRUE      TRUE
#> SQLite             FALSE      TRUE
#> ODBC               FALSE      TRUE
#> WAsP               FALSE      TRUE
#> PGeo               FALSE      TRUE
#> MSSQLSpatial       FALSE      TRUE
#> OGR_OGDI           FALSE      TRUE
#> PostgreSQL         FALSE      TRUE
#> OpenFileGDB        FALSE      TRUE
#> XPlane             FALSE      TRUE
#> DXF                FALSE      TRUE
#> Geoconcept         FALSE      TRUE
#> GeoRSS             FALSE      TRUE
#> GPSTrackMaker      FALSE      TRUE
#> VFK                FALSE      TRUE
#> PGDUMP             FALSE      TRUE
#> OSM                FALSE      TRUE
#> GPSBabel           FALSE      TRUE
#> SUA                FALSE      TRUE
#> OpenAir            FALSE      TRUE
#> OGR_PDS            FALSE      TRUE
#> WFS                FALSE      TRUE
#> HTF                FALSE      TRUE
#> AeronavFAA         FALSE      TRUE
#> Geomedia           FALSE      TRUE
#> EDIGEO             FALSE      TRUE
#> GFT                FALSE      TRUE
#> SVG                FALSE      TRUE
#> CouchDB            FALSE      TRUE
#> Cloudant           FALSE      TRUE
#> Idrisi             FALSE      TRUE
#> ARCGEN             FALSE      TRUE
#> SEGUKOOA           FALSE      TRUE
#> SEGY               FALSE      TRUE
#> XLS                FALSE      TRUE
#> ODS                FALSE      TRUE
#> XLSX               FALSE      TRUE
#> ElasticSearch      FALSE      TRUE
#> Walk               FALSE      TRUE
#> Carto              FALSE      TRUE
#> AmigoCloud         FALSE      TRUE
#> SXF                FALSE      TRUE
#> Selafin            FALSE      TRUE
#> JML                FALSE      TRUE
#> PLSCENES            TRUE      TRUE
#> CSW                FALSE      TRUE
#> VDV                FALSE      TRUE
#> TIGER              FALSE      TRUE
#> AVCBin             FALSE      TRUE
#> AVCE00             FALSE      TRUE
#> HTTP                TRUE      TRUE