st_drivers {sf}	R Documentation

Get GDAL drivers

Description

Get a list of the available GDAL drivers

Usage

st_drivers(what = "vector")

Arguments

what
character: "vector" or "raster", anything else will return all drivers.

Details

The drivers available will depend on the installation of GDAL/OGR, and can vary; the st_drivers() function shows all the drivers that are readable, and which may be written. The field vsi refers to the driver's capability to read/create datasets through the VSI*L API. See GDAL website for additional details on driver support.

Value

A data.frame with driver metadata.

Examples

st_drivers()

[Package sf version 0.7-7 Index]