get geometry from an sf object

get geometry from an sf object

# S3 method for sfc
st_geometry(obj, ...)

st_geometry(obj, ...)

# S3 method for sf
st_geometry(obj, ...)

# S3 method for sfc
st_geometry(obj, ...)

# S3 method for sfg
st_geometry(obj, ...)

st_geometry(x) <- value

Arguments

obj

object of class sf or sfc

...

ignored

x

object of class data.frame

value

object of class sfc

Value

st_geometry returns an object of class sfc, a list-column with geometries

st_geometry returns an object of class sfc. Assigning geometry to a data.frame creates an sf object, assigning it to an sf object replaces the geometry list-column.

Examples

df = data.frame(a = 1:2)
sfc = st_sfc(st_point(c(3,4)), st_point(c(10,11)))
st_geometry(sfc)

#> Geometry set for 2 features 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 3 ymin: 4 xmax: 10 ymax: 11
#> epsg (SRID):    NA
#> proj4string:    NA
#> POINT(3 4) 
#> POINT(10 11) 
#> 

st_geometry(df) <- sfc
class(df)

#> [1] "sf"         "data.frame"
#> 

st_geometry(df)

#> Geometry set for 2 features 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 3 ymin: 4 xmax: 10 ymax: 11
#> epsg (SRID):    NA
#> proj4string:    NA
#> POINT(3 4) 
#> POINT(10 11) 
#> 

st_geometry(df) <- sfc # replaces