st_geometry_type {sf}	R Documentation

Return geometry type of an object

Description

Return geometry type of an object, as a factor

Usage

st_geometry_type(x)

Arguments

x
object of class sf or sfc

Value

a factor with the geometry type of each simple feature in x

[Package sf version 0.6-2 Index]