return geometry type of an object

return geometry type of an object, as a factor

st_geometry_type(x)

Arguments

x

object of class sf or sfc

Value

returns a factor with the geometry type of each simple feature in x