spatial left or inner join

spatial left or inner join

st_join(x, y, join = st_intersects, FUN, suffix = c(".x", ".y"),
  prepared = TRUE, left = TRUE)

Arguments

x
object of class sf

y
object of class sf

join
geometry predicate function with the same profile as st_intersects; see details

FUN
aggregation function, see aggregate; in case of multiple matches, if FUN is defined, attributes of y will be aggregated using FUN; else, all combinations of x and y are returned.

suffix
length 2 character vector; see merge

prepared
logical; see st_intersects

left
logical; if TRUE carry out left join, else inner join; see also left_join

Value

an object of class sf, joined based on geometry

Details

alternative values for argument join are: st_disjoint st_touches st_crosses st_within st_contains st_overlaps st_covers st_covered_by st_equals or st_equals_exact, or user-defined functions of the same profile

Examples

a = st_sf(a = 1:3,
 geom = st_sfc(st_point(c(1,1)), st_point(c(2,2)), st_point(c(3,3))))
b = st_sf(a = 11:14,
 geom = st_sfc(st_point(c(10,10)), st_point(c(2,2)), st_point(c(2,2)), st_point(c(3,3))))
st_join(a, b)

#> Simple feature collection with 4 features and 2 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1 ymin: 1 xmax: 3 ymax: 3
#> epsg (SRID):    NA
#> proj4string:    NA
#>     a.x a.y       geom
#> 1     1  NA POINT(1 1)
#> 2     2  12 POINT(2 2)
#> 2.1   2  13 POINT(2 2)
#> 3     3  14 POINT(3 3)

st_join(a, b, left = FALSE)

#> Simple feature collection with 3 features and 2 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 2 ymin: 2 xmax: 3 ymax: 3
#> epsg (SRID):    NA
#> proj4string:    NA
#>     a.x a.y       geom
#> 2     2  12 POINT(2 2)
#> 2.1   2  13 POINT(2 2)
#> 3     3  14 POINT(3 3)

st_join(a, b, FUN = mean)

#> Simple feature collection with 3 features and 3 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 1 ymin: 1 xmax: 3 ymax: 3
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a.x Group.1  a.y       geom
#> 1   1      NA   NA POINT(1 1)
#> 2   2       2 12.5 POINT(2 2)
#> 3   3       3 14.0 POINT(3 3)

st_join(a, b, FUN = mean, left = FALSE)

#> Simple feature collection with 2 features and 3 fields
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: 2 ymin: 2 xmax: 3 ymax: 3
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a.x Group.1  a.y       geom
#> 2   2       2 12.5 POINT(2 2)
#> 3   3       3 14.0 POINT(3 3)