st_layers {sf}	R Documentation

List layers in a datasource

Description

List layers in a datasource

Usage

st_layers(dsn, options = character(0), do_count = FALSE)

Arguments

dsn
data source name (interpretation varies by driver - for some drivers, dsn is a file name, but may also be a folder, or contain the name and access credentials of a database)

options
character; driver dependent dataset open options, multiple options supported.

do_count
logical; if TRUE, count the features by reading them, even if their count is not reported by the driver

[Package sf version 0.6-2 Index]