Return 'm' range of a simple feature or simple feature set

Return 'm' range of a simple feature or simple feature set

Usage

# S3 method for m_range
is.na(x)

st_m_range(obj, ...)

# S3 method for POINT
st_m_range(obj, ...)

# S3 method for MULTIPOINT
st_m_range(obj, ...)

# S3 method for LINESTRING
st_m_range(obj, ...)

# S3 method for POLYGON
st_m_range(obj, ...)

# S3 method for MULTILINESTRING
st_m_range(obj, ...)

# S3 method for MULTIPOLYGON
st_m_range(obj, ...)

# S3 method for GEOMETRYCOLLECTION
st_m_range(obj, ...)

# S3 method for MULTISURFACE
st_m_range(obj, ...)

# S3 method for MULTICURVE
st_m_range(obj, ...)

# S3 method for CURVEPOLYGON
st_m_range(obj, ...)

# S3 method for COMPOUNDCURVE
st_m_range(obj, ...)

# S3 method for POLYHEDRALSURFACE
st_m_range(obj, ...)

# S3 method for TIN
st_m_range(obj, ...)

# S3 method for TRIANGLE
st_m_range(obj, ...)

# S3 method for CIRCULARSTRING
st_m_range(obj, ...)

# S3 method for sfc
st_m_range(obj, ...)

# S3 method for sf
st_m_range(obj, ...)

# S3 method for numeric
st_m_range(obj, ..., crs = NA_crs_)

NA_m_range_

Format

An object of class m_range of length 2.

Arguments

x

object of class m_range

obj

object to compute the m range from

...

ignored

crs

object of class crs, or argument to st_crs, specifying the CRS of this bounding box.

Value

a numeric vector of length two, with mmin and mmax values; if obj is of class sf or sfc the object if obj is of class sf or sfc the object returned has a class m_range

Details

NA_m_range_ represents the missing value for a m_range object

Examples

a = st_sf(a = 1:2, geom = st_sfc(st_point(0:3), st_point(1:4)), crs = 4326)
st_m_range(a)
#> mmin mmax 
#>    3    4 
st_m_range(c(mmin = 16.1, mmax = 16.6), crs = st_crs(4326))
#> mmin mmax 
#> 16.1 16.6