Transform or convert coordinates of simple feature

Transform or convert coordinates of simple feature

st_transform(x, crs)

# S3 method for sfc
st_transform(x, crs, ...)

# S3 method for sf
st_transform(x, crs, ...)

# S3 method for sfg
st_transform(x, crs, ...)

Arguments

x

object of class sf, sfc or sfg

crs

coordinate reference system: integer with the epsg code, or character with proj4string

...

ignored

Details

transforms coordinates of object to new projection

the st_transform method for sfg objects assumes that the crs of the object is available as an attribute of that name.

Examples

p1 = st_point(c(7,52))
p2 = st_point(c(-30,20))
sfc = st_sfc(p1, p2, crs = "+init=epsg:4326")
st_transform(sfc, "+init=epsg:3857")

#> Geometry set for 2 features 
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -3339585 ymin: 2273031 xmax: 779236.4 ymax: 6800125
#> epsg (SRID):    3857
#> proj4string:    +proj=merc +lon_0=0 +lat_ts=0 +x_0=0 +y_0=0 +a=6378137 +b=6378137 +nadgrids=@null +units=m +no_defs
#> POINT(779236.435552915 6800125.45439731) 
#> POINT(-3339584.72379821 2273030.92698769) 

st_transform(st_sf(a=2:1, geom=sfc), "+init=epsg:3857")

#> Simple feature collection with 2 features and 1 field
#> geometry type:  POINT
#> dimension:      XY
#> bbox:           xmin: -3339585 ymin: 2273031 xmax: 779236.4 ymax: 6800125
#> epsg (SRID):    3857
#> proj4string:    +proj=merc +lon_0=0 +lat_ts=0 +x_0=0 +y_0=0 +a=6378137 +b=6378137 +nadgrids=@null +units=m +no_defs
#>   a                           geom
#> 1 2 POINT(779236.435552915 6800...
#> 2 1 POINT(-3339584.72379821 227...

# FIXME: do we really want this?
st_transform(structure(p1, proj4string = "+init=epsg:4326"), "+init=epsg:3857")

#> POINT(779236.435552915 6800125.45439731)