write simple features object to file or database

write simple features object to file or database

st_write(obj, dsn, layer, driver = "ESRI Shapefile", ...,
  dataset_options = NULL, layer_options = NULL, quiet = FALSE,
  factorsAsCharacter = TRUE)

Arguments

obj

object of class sf or sfc

dsn

data source name (interpretation varies by driver - for some drivers, dsn is a file name, but may also be a folder)

layer

layer name (varies by driver, may be a file name without extension)

driver

character; OGR driver name to be used

...

ignored

dataset_options

character; driver dependent dataset creation options; multiple options supported.

layer_options

character; driver dependent layer creation options; multiple options supported.

quiet

logical; suppress info on name, driver, size and spatial reference

factorsAsCharacter

logical; convert factor objects into character strings (default), else into numbers by as.numeric.

Details

columns (variables) of a class not supported are dropped with a warning.

Examples

if (Sys.getenv("USER") %in% c("edzer", "travis")) { # load meuse to postgis
 library(sp)
 example(meuse, ask = FALSE, echo = FALSE)
 st_write(st_as_sf(meuse), "PG:dbname=postgis", "meuse_sf", driver = "PostgreSQL",
   layer_options = c("OVERWRITE=yes", "LAUNDER=true"))
 demo(nc, ask = FALSE)
 st_write(nc, "PG:dbname=postgis", "sids", driver = "PostgreSQL",
   layer_options = "OVERWRITE=true")
}

#> Writing layer meuse_sf to data source PG:dbname=postgis using driver "PostgreSQL"
#> options:        OVERWRITE=yes LAUNDER=true 
#> features:       155
#> fields:         12
#> geometry type:  POINT
#> 
#> 
#> 	demo(nc)
#> 	---- ~~
#> 
#> > ## this object was created as follows:
#> > library(sf)
#> 
#> > # nc = st_read(system.file("shapes/", package="maptools"), "sids")
#> > # st_crs(nc) = 4267 # "+proj=longlat +ellps=clrk66" or "+proj=longlat +datum=NAD27"
#> > # print(nc, n = 3)
#> > # st_write(nc, "nc.gpkg", "nc.gpkg", driver = "GPKG")
#> > 
#> > # description of the dataset, see vignette in package spdep:
#> > # https://cran.r-project.org/web/packages/spdep/vignettes/sids.pdf
#> > 
#> > datasource = { if ("GPKG" %in% st_drivers()$name)
#> + 	system.file("gpkg/nc.gpkg", package="sf")
#> + else
#> + 	system.file("shape/nc.shp", package="sf")
#> + }
#> 
#> > nc = st_read(datasource, crs = 4267,
#> + 	relation_to_geometry = c(AREA = "lattice", PERIMETER = "lattice", CNTY_ = "entity",
#> + 		CNTY_ID = "entity", NAME = "entity", FIPS = "entity", FIPSNO = "entity",
#> + 		CRESS_ID = "entity", BIR74 = "lattice", SID74 = "lattice", NWBIR74 = "lattice",
#> + 		BIR79 = "lattice", SID79 = "lattice", NWBIR79  = "lattice"))
#> Reading layer `nc.gpkg' from data source `/home/edzer/git/sfr/inst/gpkg/nc.gpkg' using driver `GPKG'
#> features:       100
#> fields:         14
#> proj4string:    +proj=longlat +datum=NAD27 +no_defs 
#> 
#> > rm(datasource)
#> Writing layer sids to data source PG:dbname=postgis using driver "PostgreSQL"
#> options:        OVERWRITE=true 
#> features:       100
#> fields:         14
#> geometry type:  MULTIPOLYGON
#> 

nc = st_read(system.file("shape/nc.shp", package="sf"), "nc", crs = 4267)

#> Reading layer `nc' from data source `/home/edzer/git/sfr/inst/shape/nc.shp' using driver `ESRI Shapefile'
#> features:       100
#> fields:         14
#> converted into: MULTIPOLYGON
#> proj4string:    +proj=longlat +datum=NAD27 +no_defs 
#> 

st_write(nc, ".", "nc")

#> Writing layer nc to data source . using driver "ESRI Shapefile"
#> 

Warning message:
GDAL Error 1: Layer 'nc' already exists

#> Creating layer nc failed.
#> 
Error in eval(substitute(expr), envir, enclos): Layer creation failed.