stars {sf}	R Documentation

functions only exported to be used internally by stars

Description

functions only exported to be used internally by stars

Usage

.get_layout(bb, n, total_size, key.pos, key.size)

.degAxis(side, at, labels, ..., lon, lat, ndiscr, reset)

.image_scale(z, col, breaks = NULL, key.pos, add.axis = TRUE, at = NULL,
  ..., axes = FALSE)

.image_scale_factor(z, col, breaks = NULL, key.pos, add.axis = TRUE,
  at = NULL, ..., axes = FALSE, key.size)

Arguments

bb
ignore

n
ignore

total_size
ignore

key.pos
ignore

key.size
ignore

side
ignore

at
ignore

labels
ignore

...
ignore

lon
ignore

lat
ignore

ndiscr
ignore

reset
ignore

z
ignore

col
ignore

breaks
ignore

add.axis
ignore

axes
ignore

[Package sf version 0.6-2 Index]