functions only exported to be used internally by stars

functions only exported to be used internally by stars

Usage

.get_layout(bb, n, total_size, key.pos, key.width, mfrow = NULL, main = NULL)

.degAxis(side, at, labels, ..., lon, lat, ndiscr, reset)

.image_scale(
  z,
  col,
  breaks = NULL,
  key.pos,
  add.axis = TRUE,
  at = NULL,
  ...,
  axes = FALSE,
  key.length,
  logz = FALSE,
  lab = "",
  cex.axis = par("cex.axis")
)

.image_scale_factor(
  z,
  col,
  key.pos,
  add.axis = TRUE,
  ...,
  axes = FALSE,
  key.width,
  key.length,
  cex.axis = par("cex.axis")
)

Arguments

bb

ignore

n

ignore

total_size

ignore

key.pos

ignore

key.width

ignore

mfrow

length-2 integer vector with number of rows, columns

main

main or sub title

side

ignore

at

ignore

labels

ignore

...

ignore

lon

ignore

lat

ignore

ndiscr

ignore

reset

ignore

z

ignore

col

ignore

breaks

ignore

add.axis

ignore

axes

ignore

key.length

ignore

logz

ignore

lab

ignore

cex.axis

see par